Rename TransactionConsumptionAction to TransactionConsumptionActionParams and update README.md
